requirements.txt
aiogram>=3.0.0
pandas>=2.0.0
openpyxl>=3.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.1
asyncio>=3.4.3
vk-api>=11.9.9
telethon>=1.24.0
yadisk>=1.3.0
schedule>=1.2.0
python-telegram-bot>=20.0
aiosqlite>=0.17.0
flask>=2.0.0
pyngrok>=6.0.0
openai>=1.35.0
yadisk-async>=1.4.0
cryptg>=0.2.post4
pillow>=8.3.2
requests>=2.26.0
python-dateutil>=2.8.2
psycopg2-binary>=2.9.1 